abstraction for consistency-preserving event DB
The event DB is a fairly clean abstraction. It helps solve two major problems:
 * Keeping all clients up to date on their relevant state.
 * Dealing with user references and forceful logouts.

In addition, the event DB can handle buffering and event dropping, allowing us to remove some load from the ugly Connection abstraction. Also, by abstracting all of this in the DB layer, it'll be easier to scale the application later.